+++ Changes.md
@@ -28,6 +28,7 @@ Fixes
 - ViewportGadget :
   - Fixed `setCenterOfInterest()` so that it doesn't emit `cameraChangedSignal()` if the center of interest is unchanged.
   - Added GIL release in `setViewport()` Python bindings.
+- ArnoldRender : Fixed rendering of single-channel AOVs specified using Gaffer's generic `float|int|uint <name>` syntax. Outputs specified using Arnold's `<name> FLOAT|INT|UINT` syntax will now issue a warning, and should be updated to use the generic syntax.
 
 API
 ---